+++ rdsai-20250507/alibabacloud_rdsai20250507/models/_get_scheduled_reports_response_body.py
@@ -102,7 +102,9 @@ class GetScheduledReportsResponseBodyReports(DaraModel):
         self,
         created_time: str = None,
         end_time: str = None,
+        region_id: str = None,
         report_language: str = None,
+        report_type: str = None,
         start_time: str = None,
         status: str = None,
         task_id: str = None,
@@ -111,7 +113,9 @@ class GetScheduledReportsResponseBodyReports(DaraModel):
         self.created_time = created_time
         # The end time of the inspection task. The time follows the ISO 8601 standard in the YYYY-MM-DDTHH:mm:ssZ format.
         self.end_time = end_time
+        self.region_id = region_id
         self.report_language = report_language
+        self.report_type = report_type
         # The start time of the inspection task. The time follows the ISO 8601 standard in the YYYY-MM-DDTHH:mm:ssZ format.
         self.start_time = start_time
         # The status of the task.
@@ -133,8 +137,14 @@ class GetScheduledReportsResponseBodyReports(DaraModel):
         if self.end_time is not None:
             result['EndTime'] = self.end_time
 
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+
         if self.report_language is not None:
             result['ReportLanguage'] = self.report_language
+
+        if self.report_type is not None:
+            result['ReportType'] = self.report_type
 
         if self.start_time is not None:
             result['StartTime'] = self.start_time
@@ -155,8 +165,14 @@ class GetScheduledReportsResponseBodyReports(DaraModel):
         if m.get('EndTime') is not None:
             self.end_time = m.get('EndTime')
 
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+
         if m.get('ReportLanguage') is not None:
             self.report_language = m.get('ReportLanguage')
+
+        if m.get('ReportType') is not None:
+            self.report_type = m.get('ReportType')
 
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')