+++ rdsai-20250507/alibabacloud_rdsai20250507/models/_get_stand_alone_reports_request.py
@@ -10,6 +10,7 @@ class GetStandAloneReportsRequest(DaraModel):
         end_time: str = None,
         page_number: int = None,
         page_size: int = None,
+        report_type: str = None,
         start_time: str = None,
     ):
         # The task end time based on which the reports are filtered. If you do not specify this parameter, all reports are returned.
@@ -18,6 +19,7 @@ class GetStandAloneReportsRequest(DaraModel):
         self.page_number = page_number
         # The number of records on each page. Default value: 20. Maximum value: 100.
         self.page_size = page_size
+        self.report_type = report_type
         # The task start time based on which the reports are filtered. If you do not specify this parameter, all reports are returned.
         self.start_time = start_time
 
@@ -38,6 +40,9 @@ class GetStandAloneReportsRequest(DaraModel):
         if self.page_size is not None:
             result['PageSize'] = self.page_size
 
+        if self.report_type is not None:
+            result['ReportType'] = self.report_type
+
         if self.start_time is not None:
             result['StartTime'] = self.start_time
 
@@ -53,6 +58,9 @@ class GetStandAloneReportsRequest(DaraModel):
 
         if m.get('PageSize') is not None:
             self.page_size = m.get('PageSize')
+
+        if m.get('ReportType') is not None:
+            self.report_type = m.get('ReportType')
 
         if m.get('StartTime') is not None:
             self.start_time = m.get('StartTime')